test(iceberg): Revamp Iceberg tests
